Add append algorithm result in problem file, large number scientific notation format in plot

+++ src/main.py
@@ -8,13 +8,22 @@ outputFolder = '../output/'
 
 def saveResult(result, problem, algorithmName):
     fileName = outputFolder + problem + '.out'
-    os.makedirs(os.path.dirname(fileName), exist_ok=True)
-    with open(fileName, 'w') as f:
-        print('writting', problem, algorithmName)
-        f.write(str(s.optimum) + ',')
-        f.write(algorithmName + ',')
-        result = ','.join(map(str, result)) 
-        f.write(result)
+    if os.path.exists(fileName) :
+        with open(fileName, 'a') as f:
+            print('writting', problem, algorithmName)
+            f.write('\n')
+            f.write(str(s.optimum) + ',')
+            f.write(algorithmName + ',')
+            result = ','.join(map(str, result)) 
+            f.write(result)
+    else:
+        os.makedirs(os.path.dirname(fileName), exist_ok=True)
+        with open(fileName, 'w') as f:
+            print('writting', problem, algorithmName)
+            f.write(str(s.optimum) + ',')
+            f.write(algorithmName + ',')
+            result = ','.join(map(str, result)) 
+            f.write(result)
 
 def plotComparison(problems):
     for problem in problems:
@@ -28,7 +37,7 @@ def plotComparison(problems):
                 opt = float(splittedLine[0])
                 algorithmName = str(splittedLine[1])
                 splittedResults = list(map(lambda x: float(x), splittedLine[2:]))
-                plt.loglog(range(len(splittedResults)), splittedResults, label=algorithmName)
+                plt.plot(range(len(splittedResults)), splittedResults, label=algorithmName)
         
         plt.axhline(y=opt, linestyle='dashed', color='c')
         plt.legend()
@@ -52,8 +61,8 @@ if __name__ == '__main__':
             # name, out = s.randomWalk(0.5)
             # name, out = s.metropolisHasting()
             # name, out = s.hillClimbing()
-            name, out = s.simulatedAnnealing(10**3, 10**(-8), s.linearCoolingStrategy, 0.99)
-            # name, out = s.simulatedAnnealing(10**30, 10**(-8), s.expCoolingStrategy, 0.99)
+            # name, out = s.simulatedAnnealing(10**3, 10**(-8), s.linearCoolingStrategy, 0.99)
+            name, out = s.simulatedAnnealing(10**30, 10**(-8), s.expCoolingStrategy, 0.99)
             # name, out = s.simulatedAnnealing(10**5, 10**(-8), s.dynamicCoolingStrategy, 0.5)
             
             if s.bestSolution.v > bestFound:
